Module Std.Group

Group is a non-empty set that is a result of partitioning of an underlying set S into a set of non-intersecting and non-empty subsets that cover set S. See Partition for more information.

type 'a t = 'a group

val enum : 'a group -> 'a Regular.Std.seq

enum group enumerates all elements of a group, including the designated one.

val mem : 'a group -> 'a -> bool

mem group x checks membership of x in a given group.

val top : 'a group -> 'a

top group returns the top element of a group also known as a representative element. The function is total since groups is guaranteed to be non-empty.

val to_equiv : 'a group -> equiv

to_equiv g returns the ordinal number representing the particular group g

val pp : 'a Regular.Std.printer -> 'a t Regular.Std.printer

pp pp_elem g prints group g using element printer pp_elem