Module Std.Frontier

Frontier maps each node into a possibly empty set of nodes. This is used for representing dominance and post-dominance frontiers.

type 'a t = 'a frontier

val enum : 'a t -> 'a -> 'a Regular.Std.seq

enum f x enumerates frontier of x

val mem : 'a t -> 'a -> 'a -> bool

mem f x y is true if y is in a frontier of x

val to_sequence : 'a t -> 'a Regular.Std.seq

to_sequence frontier enumerates all elements of a frontier

val pp : 'a Regular.Std.printer -> 'a t Regular.Std.printer

pp pp_elt instantiates a pretty-printer for a given element. See Tree.pp for more information.