Module 1-G.Edge

Graph edges

type t = edge

type node = node

type graph = t

type label

val create : node -> node -> label -> t

create x y l creates an edge connecting nodes x and y labeled with a a given label l

val label : t -> label

label e returns a label of an edge e

val src : t -> node

src e returns a source of an edge e

val dst : t -> node

dst e returns a destination of an edge e

val mem : t -> graph -> bool

mem e g is true if e ∈ E.

val insert : t -> graph -> graph

insert e g returns a graph g' with a set of edges extended with edge e. If src e or dst e wasn't in the set of nodes N, then it is extended as well, so that axioms of graph are preserved.

Postconditions:

          - E(g') = E(g) ∪ {e}.

val update : t -> label -> graph -> graph

update e l g if edge e exists in graph g then return a new graph g' in which edge e is associated with label l. Otherwise return g unchanged.

Postcondition:

          - E(g) ⊆ E(g')
          - N(g) ⊆ N(g')
          - e ∉ E(g) → e ∉ E(g').
          - e ∈ E(g) → ε(g')e = l.

val remove : t -> graph -> graph

remove e g returns a graph g' that doesn't contain edge e.

Postconditions:

          - E(g') = E(g) \ {e}.

include Regular.Std.Opaque.S with type t := t

type t

include Core_kernel.Comparable.S with type Edge.t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : (t, comparator_witness) Base__.Comparator.comparator

val validate_lbound : min:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_ubound : max:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_bound : min:t Base__.Maybe_bound.t -> max:t Base__.Maybe_bound.t -> t Base__.Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : sig ... end

module Set : sig ... end

include Core_kernel.Hashable.S with type Edge.t := t

val compare : t -> t -> Core_kernel__.Import.int

val hash_fold_t : Ppx_hash_lib.Std.Hash.state -> t -> Ppx_hash_lib.Std.Hash.state

val hash : t -> Ppx_hash_lib.Std.Hash.hash_value

val hashable : t Core_kernel__.Hashtbl.Hashable.t

module Table : sig ... end

module Hash_set : sig ... end

module Hash_queue : sig ... end