Module Std.Path

Path between two nodes.

path is a walk without repetitions

type 'e t = 'e path

representation type

val start : 'e t -> 'e

start p the starting edge of a path p

val finish : 'e t -> 'e

finish p the last edge of a path p

val edges : 'e t -> 'e Regular.Std.seq

edges p a sequence of edges from start to finish

val edges_rev : 'e t -> 'e Regular.Std.seq

edges_rev p a reversed sequence from finish to start

val weight : 'e t -> int

weight p total weight of a path

val length : 'e t -> int

amount of edges in a path

val pp : 'a Regular.Std.printer -> 'a t Regular.Std.printer

pp pp_elt constructs a pretty based on element printer pp_elt